FEC-3643 comment Changed the int values of the progress to fload for accuration

+++ playerSDK/src/main/java/com/kaltura/playersdk/players/KPlayerController.java
@@ -143,7 +143,7 @@ public class KPlayerController implements KPlayerListener{
         return this.player.getCurrentPlaybackTime();
     }
 
-    public void setCurrentPlaybackTime(int currentPlaybackTime) {
+    public void setCurrentPlaybackTime(float currentPlaybackTime) {
         this.player.setCurrentPlaybackTime(currentPlaybackTime);
     }
